Trying to tidy up the process of saving state machine models and make it more thorough.
